requirements.txt
absl-py==0.13.0
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1610094673755/work
astunparse==1.6.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1631636467320/work
chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1610093613037/work
charset-normalizer==2.0.5
chart-studio==1.1.0
clang==5.0
click==8.0.1
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1616851523607/work
cycler==0.10.0
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1627075104865/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
imageio==2.3.0
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1631291142144/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1630172558090/work
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1610146808808/work
Jinja2==3.0.1
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1612898609989/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1630336909568/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1612125283939/work
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1630173773255/work
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1628881755815/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
numpy==1.19.5
oauthlib==3.1.1
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
patsy==0.5.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1630696651849/work
plotly @ file:///Users/nicolas/opt/miniconda3/envs/env/conda-bld/plotly_1630459557113/work/packages/python/plotly/dist/plotly-5.3.1.tar.gz
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1629903925368/work
protobuf==3.18.0
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1610291470465/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1628233184753/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1626393743643/work
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1630910586219/work
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1628206569514/work
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1629095986539/work
six==1.15.0
statsmodels @ file:///Users/runner/miniforge3/conda-bld/statsmodels_1630699970101/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1626090218611/work
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1626092076920/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1625488906146/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1630423529112/work
typing-extensions==3.7.4.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1624634538755/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==2.0.1
wrapt==1.12.1
